feat(poc): add initialization manager as resource orchestrator
